fix possible stereo chem issues and reorder the package
