+++ QUBEKit/engines/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 """
 Ensures that errors with particular engines are only raised if said engine is being used.
 No point raising "cannot execute psi4 command" if user is using g09. See try_load docstring for info.
