+++ QUBEKit/engines/base_engine.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 from typing import Dict, Optional
 
 import qcelemental as qcel
