+++ QUBEKit/lennard_jones.py
@@ -4,7 +4,7 @@ import math
 from collections import namedtuple
 from typing import Dict, List, Optional
 
-from QUBEKit.ligand import Ligand
+from QUBEKit.molecules import Ligand
 from QUBEKit.utils import constants
