+++ QUBEKit/molecules/__init__.py
@@ -1,3 +1,3 @@
-from QUBEKit.molecules.components import Atom, Element, ExtraSite
+from QUBEKit.molecules.components import Atom, Bond, Element, ExtraSite
 from QUBEKit.molecules.ligand import Ligand
 from QUBEKit.molecules.protein import Protein
