+++ QUBEKit/parametrisation/__init__.py
@@ -1,10 +1,5 @@
-#!/usr/bin/env python3
-
-from QUBEKit.utils.helpers import try_load
-
-from .base_parametrisation import Parametrisation
-from .openff import OpenFF
-from .xml import XML
-from .xml_protein import XMLProtein
-
-AnteChamber = try_load("AnteChamber", "QUBEKit.parametrisation.antechamber")
+from QUBEKit.parametrisation.antechamber import AnteChamber
+from QUBEKit.parametrisation.base_parametrisation import Parametrisation
+from QUBEKit.parametrisation.openff import OpenFF
+from QUBEKit.parametrisation.xml import XML
+from QUBEKit.parametrisation.xml_protein import XMLProtein
