+++ QUBEKit/parametrisation/base_parametrisation.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import xml.etree.ElementTree as ET
 from collections import OrderedDict
 from copy import deepcopy
