+++ QUBEKit/tests/protein/parametrisation_test.py
@@ -2,9 +2,6 @@ import shutil
 
 import pytest
 
-from QUBEKit.ligand import Protein
-import shutil
-
 from QUBEKit.molecules import Protein
 from QUBEKit.parametrisation import XMLProtein
 from QUBEKit.utils.file_handling import get_data
